distributed models per handler
